Split off vertex/fragment shaders for readability
